+++ src/Adapter/CurlAdapter.php
@@ -24,7 +24,6 @@ use Berlioz\Http\Message\Request;
 use Berlioz\Http\Message\Response;
 use Berlioz\Http\Message\Stream\MemoryStream;
 use Berlioz\Http\Message\Uri;
-use CurlHandle;
 use DateTimeImmutable;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
@@ -114,7 +113,7 @@ class CurlAdapter extends AbstractAdapter
         $bodyStream = new MemoryStream();
 
         // Init CURL
-        $ch = $this->initCurl(
+        $curlOptions = $this->prepareCurlOptions(
             $request,
             [
                 CURLOPT_HEADERFUNCTION => fn($ch, $data) => $headersStream->write($data),
@@ -126,6 +125,8 @@ class CurlAdapter extends AbstractAdapter
             ],
             $context,
         );
+        $ch = curl_init();
+        curl_setopt_array($ch, $curlOptions);
 
         try {
             // Execute CURL request
@@ -206,55 +207,53 @@ class CurlAdapter extends AbstractAdapter
     }
 
     /**
-     * Init CURL options.
+     * Prepare CURL options.
      *
      * @param RequestInterface $request
      * @param array $options
      * @param HttpContext|null $context
      *
-     * @return CurlHandle
+     * @return array
      */
-    protected function initCurl(
+    protected function prepareCurlOptions(
         RequestInterface $request,
         array $options = [],
         ?HttpContext $context = null,
-    ): CurlHandle {
-        // CURL init
-        $ch = curl_init();
-        curl_setopt_array($ch, $options);
+    ): array {
+        $options = array_replace($this->options, $options);
 
         // HTTP Version
         switch ($request->getProtocolVersion()) {
             case 1.0:
-                $curlOpts[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_1_0;
+                $options[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_1_0;
                 break;
             case 2.0:
-                $curlOpts[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_2_0;
+                $options[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_2_0;
                 break;
             case 1.1:
             default:
-                $curlOpts[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_1_1;
+                $options[CURLOPT_HTTP_VERSION] = CURL_HTTP_VERSION_1_1;
         }
 
         // URL of request
-        $curlOpts[CURLOPT_CUSTOMREQUEST] = $request->getMethod();
-        $curlOpts[CURLOPT_URL] = $request->getUri();
+        $options[CURLOPT_CUSTOMREQUEST] = $request->getMethod();
+        $options[CURLOPT_URL] = $request->getUri();
 
         // Timeouts
-        $curlOpts[CURLOPT_NOSIGNAL] = true;
+        $options[CURLOPT_NOSIGNAL] = true;
 
         // Headers
         {
-            $curlOpts[CURLOPT_HEADER] = false;
-            $curlOpts[CURLINFO_HEADER_OUT] = false;
-            $curlOpts[CURLOPT_HTTPHEADER] = $this->getHeadersLines($request);
+            $options[CURLOPT_HEADER] = false;
+            $options[CURLINFO_HEADER_OUT] = false;
+            $options[CURLOPT_HTTPHEADER] = $this->getHeadersLines($request);
         }
 
-        $curlOpts[CURLOPT_FOLLOWLOCATION] = false;
+        $options[CURLOPT_FOLLOWLOCATION] = false;
 
         if ($request->getBody()->getSize() > 0) {
-            $curlOpts[CURLOPT_POST] = true;
-            $curlOpts[CURLOPT_POSTFIELDS] = (string)$request->getBody();
+            $options[CURLOPT_POST] = true;
+            $options[CURLOPT_POSTFIELDS] = (string)$request->getBody();
         }
 
         // HTTP Context
@@ -284,11 +283,9 @@ class CurlAdapter extends AbstractAdapter
             $contextOptions[CURLOPT_SSLKEY] = $context->ssl_local_pk;
             $contextOptions[CURLOPT_KEYPASSWD] = $context->ssl_local_passphrase;
 
-            $curlOpts = array_replace($curlOpts, array_filter($contextOptions, fn($value) => null !== $value));
+            $options = array_replace($options, array_filter($contextOptions, fn($value) => null !== $value));
         }
 
-        curl_setopt_array($ch, $curlOpts);
-
-        return $ch;
+        return $options;
     }
 }